Add files via upload
MNIST is the de facto " hello world" dataset of computer vision. This classic data set of handwritten images has served as the basis for benchmarking  classification algorithms. As new machine learning techniques emerge, MNIST remains a reliable resource for researchers and learners alike.

In this note book our goal is to correctly identify digits from a dataset of thousands of handwritten images.

### Task Performed
1. Prepared the dataset for the model
2. Installed Tensor flow 2.x
3. Developed Single Layer Perceptron model for classifying the handwritten digits
4. Plotted the change in accuracy per epochs
5. Evaluated the model on the testing data
6. Analyzed the model summary
7. Added hidden layer to the model to make it Multi layer perceptron
8. Added dropout to prevent overfitting and checked its effects on accuracy
9. Increased the number of Hidden Layer neuron and checked its effect on accuracy
10. Used different optimizers and checked its effect on accuracy
11.Manupulated the batch size and epochs and checked its effect on accuracy